Make `ilab` model agnostic
both generate and serve already had a flag for the context size, meaning the user can override the default of 4096 in those two scenarios.

Added a new flag called `model-family` to serve which currently supports `merlinite` or `mixtral`. This is then used to load the proper chat template from a hardcoded map of supported templates. If nothing is specified, we default to the current one hardcoded in instructlab

Also, add `model-name` to `ilab train` as linux train had a hardcoded merlinite reference for the model name. `model-path` also needed to be added to generate and chat since if a server is not running when either of these are started, we default to using the model_path in the server config (merlinite by default)

Most other "hardcoded" references are really just defaults which can be overridden by the user.

Signed-off-by: Charlie Doern <cdoern@redhat.com>